backend/src/utils/requirements.txt
flask==2.3.2
flask-cors==4.0.0
sentence-transformers==2.2.2
numpy==1.24.3
faiss-cpu==1.7.4
transformers==4.31.0
torch==2.0.1